Update InfiniteArray.java to handle when index > 6
Before, when index was greater than 6, the code would show an error. So, I added a check within the while loop in the ans method to ensure that the end index does not exceed the array size. If the calculated end index is greater than or equal to the array length, I set it to the last valid index to prevent an ArrayIndexOutOfBoundsException when accessing the array.